feat: remove deprecated methods
implementing ENG-3162
