+++ indykite_sdk_node/package.json
@@ -29,12 +29,12 @@
     "@babel/core": "7.24.0",
     "@babel/preset-env": "7.24.0",
     "@babel/preset-typescript": "7.23.3",
-    "@types/argparse": "2.0.14",
+    "@types/argparse": "2.0.15",
     "@types/bcrypt": "^5.0.2",
     "@types/google-protobuf": "3.15.12",
     "@types/jest": "29.5.12",
     "@types/jsonwebtoken": "9.0.6",
-    "@types/node": "20.11.24",
+    "@types/node": "20.11.25",
     "@types/uuid": "9.0.8",
     "@typescript-eslint/eslint-plugin": "7.1.1",
     "@typescript-eslint/parser": "7.1.1",
@@ -52,7 +52,7 @@
     "rimraf": "5.0.5",
     "ts-jest": "29.1.2",
     "ts-node": "^10.9.2",
-    "typedoc": "0.25.10",
+    "typedoc": "0.25.12",
     "typescript": "5.3.3"
   },
   "dependencies": {
@@ -60,7 +60,7 @@
     "@protobuf-ts/runtime": "2.9.3",
     "@protobuf-ts/runtime-rpc": "2.9.3",
     "google-protobuf": "3.21.2",
-    "jose": "^5.2.2",
+    "jose": "^5.2.3",
     "protobufjs": "7.2.6",
     "uuid": "^9.0.1"
   },
