+++ indykite_sdk_node/src/index.ts
@@ -14,8 +14,7 @@ export * from './grpc/indykite/authorization/v1beta1/authorization_service';
 export * from './sdk/authorization';
 export * from './sdk/identity';
 export * from './sdk/config';
-export { IdentityKnowledgeClient } from './sdk/knowledge';
-export * from './sdk/knowledgev2';
+export * from './sdk/knowledge';
 export * from './sdk/http';
 export * from './sdk/ingest';
 export * from './sdk/model';
