+++ indykite_sdk_node/src/sdk/__test__/knowledge/new_client.test.ts
@@ -1,6 +1,6 @@
-import { IdentityKnowledgeAPIClient } from '../../../grpc/indykite/knowledge/v1beta1/identity_knowledge_api.grpc-client';
+import { IdentityKnowledgeAPIClient } from '../../../grpc/indykite/knowledge/v1beta2/identity_knowledge_api.grpc-client';
 import { SdkClient } from '../../client/client';
-import { IdentityKnowledgeClient } from '../../knowledge';
+import { IdentityKnowledgeReadClient } from '../../knowledge';
 import { applicationTokenMock } from '../../utils/test_utils';
 
 afterEach(() => {
@@ -9,24 +9,24 @@ afterEach(() => {
 
 describe('when a new client is created', () => {
   describe('when no error is thrown', () => {
-    let returnedValue: IdentityKnowledgeClient;
+    let returnedValue: IdentityKnowledgeReadClient;
 
     beforeEach(async () => {
       jest
         .spyOn(SdkClient, 'createIdentityInstance')
         .mockImplementation(() => Promise.resolve({ client: {} } as SdkClient));
 
-      returnedValue = await IdentityKnowledgeClient.createInstance(
+      returnedValue = await IdentityKnowledgeReadClient.createInstance(
         JSON.stringify(applicationTokenMock),
       );
     });
 
     it('creates a new instance', () => {
-      expect(SdkClient.createIdentityInstance).toBeCalledWith(
+      expect(SdkClient.createIdentityInstance).toHaveBeenCalledWith(
         IdentityKnowledgeAPIClient,
         JSON.stringify(applicationTokenMock),
       );
-      expect(returnedValue).toBeInstanceOf(IdentityKnowledgeClient);
+      expect(returnedValue).toBeInstanceOf(IdentityKnowledgeReadClient);
     });
   });
 
@@ -39,14 +39,14 @@ describe('when a new client is created', () => {
         .spyOn(SdkClient, 'createIdentityInstance')
         .mockImplementation(() => Promise.reject(error));
       try {
-        await IdentityKnowledgeClient.createInstance(JSON.stringify(applicationTokenMock));
+        await IdentityKnowledgeReadClient.createInstance(JSON.stringify(applicationTokenMock));
       } catch (err) {
         caughtError = err as Error;
       }
     });
 
     it('throws an error', () => {
-      expect(SdkClient.createIdentityInstance).toBeCalledWith(
+      expect(SdkClient.createIdentityInstance).toHaveBeenCalledWith(
         IdentityKnowledgeAPIClient,
         JSON.stringify(applicationTokenMock),
       );
