+++ indykite_sdk_node/src/sdk/identity.ts
@@ -8,40 +8,18 @@ import {
   CreateInvitationRequest,
   CreateOAuth2ConsentVerifierRequest,
   CreateOAuth2ConsentVerifierResponse,
-  DeleteDigitalTwinRequest,
-  ListDigitalTwinsRequest,
   EnrichTokenRequest,
-  GetDigitalTwinRequest,
   ListConsentsRequest,
-  PatchDigitalTwinRequest,
-  RegisterDigitalTwinWithoutCredentialRequest,
-  RegisterDigitalTwinWithoutCredentialResponse,
   ResendInvitationRequest,
   RevokeConsentRequest,
   TokenIntrospectRequest,
 } from '../grpc/indykite/identity/v1beta2/identity_management_api';
-import { DigitalTwin, DigitalTwinIdentifier } from '../grpc/indykite/identity/v1beta2/model';
 import * as sdkTypes from './model';
-import type { GetDigitalTwinResponse } from '../grpc/indykite/identity/v1beta2/identity_management_api';
 
 import { SdkErrorCode, SdkError } from './error';
 import { SdkClient } from './client/client';
 import { IdentityManagementAPIClient } from '../grpc/indykite/identity/v1beta2/identity_management_api.grpc-client';
-import {
-  BatchOperationResult,
-  PropertyBatchOperation,
-  PropertyFilter,
-  PropertyMask,
-} from '../grpc/indykite/identity/v1beta2/attributes';
 import { Utils } from './utils/utils';
-import { PatchResult, Property } from './model';
-import {
-  ImportDigitalTwin,
-  ImportDigitalTwinsRequest,
-  ImportDigitalTwinsResponse,
-} from '../grpc/indykite/identity/v1beta2/import';
-import { Readable } from 'stream';
-import { IndexFixer, streamSplitter } from './utils/stream';
 
 /**
  * @category Clients
@@ -103,396 +81,6 @@ export class IdentityClient {
           resolve(response);
         } catch (err) {
           reject(err);
-        }
-      });
-    });
-  }
-
-  /**
-   * Receive all properties for given digital twin.
-   * @since 0.4.1
-   * @example
-   * async function example(sdk: IdentityClient) {
-   *   const dt = await sdk.getDigitalTwin(
-   *     DigitalTwin.create({
-   *       id: DIGITAL_TWIN_ID,
-   *       tenantId: TENANT_ID,
-   *     }),
-   *     [
-   *       grpcIdentityAttributes.PropertyMask.create({
-   *         definition: {
-   *           property: 'email',
-   *         },
-   *       }),
-   *     ],
-   *   );
-   *   console.log(JSON.stringify(dt, null, 2));
-   * }
-   */
-  getDigitalTwin(
-    digitalTwin: DigitalTwin,
-    properties?: PropertyMask[],
-  ): Promise<GetDigitalTwinResponse> {
-    return this.processGetDigitalTwinRequest(
-      {
-        filter: {
-          oneofKind: 'digitalTwin',
-          digitalTwin,
-        },
-      },
-      properties,
-    );
-  }
-
-  /**
-   * Receive all properties for given digital twin.
-   * @since O.4.1
-   * @example
-   * async function example(sdk: IdentityClient) {
-   *   const dt = await sdk.getDigitalTwinByToken(
-   *     USER_ACCESS_TOKEN,
-   *     [
-   *       grpcIdentityAttributes.PropertyMask.create({
-   *         definition: {
-   *           property: 'email',
-   *         },
-   *       }),
-   *     ],
-   *   );
-   *   console.log(JSON.stringify(dt, null, 2));
-   * }
-   */
-  getDigitalTwinByToken(
-    token: string,
-    properties?: PropertyMask[],
-  ): Promise<GetDigitalTwinResponse> {
-    return this.processGetDigitalTwinRequest(
-      {
-        filter: {
-          oneofKind: 'accessToken',
-          accessToken: token,
-        },
-      },
-      properties,
-    );
-  }
-
-  /**
-   * **
-   * Receive all properties for given digital twin.
-   * @since 0.4.1
-   * @example
-   * async function example(sdk: IdentityClient) {
-   *   const dt = await sdk.getDigitalTwinByProperty(
-   *     grpcIdentityAttributes.PropertyFilter.create({
-   *       tenantId: TENANT_ID,
-   *       type: 'email',
-   *       value: grpcStruct.Value.fromJson({ stringValue: 'user@example.com' }),
-   *     }),
-   *     [
-   *       grpcIdentityAttributes.PropertyMask.create({
-   *         definition: {
-   *           property: 'email',
-   *         },
-   *       }),
-   *     ],
-   *   );
-   *   console.log(JSON.stringify(dt, null, 2));
-   * }
-   */
-  getDigitalTwinByProperty(
-    propertyFilter: PropertyFilter,
-    properties?: PropertyMask[],
-  ): Promise<GetDigitalTwinResponse> {
-    return this.processGetDigitalTwinRequest(
-      {
-        filter: {
-          oneofKind: 'propertyFilter',
-          propertyFilter,
-        },
-      },
-      properties,
-    );
-  }
-
-  /**
-   * This function lists DigitalTwins matching the filter.
-   * @param tenantId The Tenant ID.
-   * @param collectionId CollectionId, relative to `parent`, to list.
-   * @param properties The requested property values.
-   * @param pageSize The maximum number of documents to return.
-   * @param pageToken The `nextPageToken` value returned from a previous List request.
-   * @param orderBy Expression to sort results by. For example: `priority desc, name`.
-   */
-  listDigitalTwins(
-    tenantId: string,
-    collectionId: string,
-    properties: string[],
-    pageSize?: number,
-    pageToken?: string,
-    orderBy?: string,
-  ): Promise<{ digitalTwins: sdkTypes.DigitalTwin[]; nextPageToken?: string }> {
-    return new Promise((resolve, reject) => {
-      const request = ListDigitalTwinsRequest.create({
-        tenantId,
-        properties: Property.fromPropertiesList(properties),
-        collectionId,
-      });
-
-      if (pageToken !== undefined) request.pageToken = pageToken;
-      if (orderBy !== undefined) request.orderBy = orderBy;
-      if (pageSize !== undefined) request.pageSize = pageSize;
-
-      this.client.listDigitalTwins(request, (err, res) => {
-        if (err) reject(err);
-        else if (!res) resolve({ digitalTwins: [] });
-        else
-          resolve({
-            digitalTwins: res.digitalTwin.map((result) =>
-              sdkTypes.DigitalTwin.deserialize({ digitalTwin: result }),
-            ),
-            nextPageToken: res.nextPageToken,
-          });
-      });
-    });
-  }
-
-  private patch(request: PatchDigitalTwinRequest): Promise<PatchResult[]> {
-    return new Promise((resolve, reject) => {
-      this.client.patchDigitalTwin(request, (err, response) => {
-        if (err) reject(err);
-        else {
-          if (response && response.result) {
-            resolve(response.result.map((result) => PatchResult.deserialize(result)));
-          } else {
-            reject(new SdkError(SdkErrorCode.SDK_CODE_1, 'Missing patch response'));
-          }
-        }
-      });
-    });
-  }
-
-  patchProperties(
-    digitalTwinId: string,
-    tenantId: string,
-    dt: sdkTypes.DigitalTwin,
-    forceDelete = false,
-  ): Promise<PatchResult[]> {
-    const patchdt = DigitalTwin.fromJson({});
-    patchdt.id = digitalTwinId;
-    patchdt.tenantId = tenantId;
-
-    const dti = DigitalTwinIdentifier.fromJson({
-      digitalTwin: DigitalTwin.toJson(patchdt),
-    });
-
-    const request = PatchDigitalTwinRequest.fromJson({
-      id: DigitalTwinIdentifier.toJson(dti),
-      forceDelete,
-    });
-    request.operations = dt.getPatchOperationsAndReset();
-    return this.patch(request);
-  }
-
-  patchPropertiesByToken(
-    token: string,
-    dt: sdkTypes.DigitalTwin,
-    forceDelete = false,
-  ): Promise<PatchResult[]> {
-    if (token.length < 32) {
-      throw new Error('Token must be 32 chars or more.');
-    }
-    const request = PatchDigitalTwinRequest.fromJson({
-      id: Utils.createDigitalTwinIdFromToken(token),
-      forceDelete,
-    });
-    request.operations = dt.getPatchOperationsAndReset();
-    return this.patch(request);
-  }
-
-  /**
-   * @since 0.4.1
-   * @example
-   * // Add a new 'email' property with 'user@example.com' value
-   * async function example(sdk: IdentityClient, dt: DigitalTwin) {
-   *   const opResult = await sdk.patchDigitalTwin(dt, [
-   *     {
-   *       operation: {
-   *         oneofKind: 'add',
-   *         add: {
-   *           id: '',
-   *           definition: {
-   *             property: 'email',
-   *             context: '',
-   *             type: '',
-   *           },
-   *           value: {
-   *             oneofKind: 'objectValue',
-   *             objectValue: {
-   *               value: { oneofKind: 'stringValue', stringValue: 'user@example.com' },
-   *             },
-   *           },
-   *         },
-   *       },
-   *     },
-   *   ]);
-   *   console.log(JSON.stringify(opResult, null, 2));
-   * }
-   */
-  patchDigitalTwin(
-    digitalTwin: DigitalTwin,
-    operations: PropertyBatchOperation[],
-    forceDelete = false,
-  ): Promise<BatchOperationResult> {
-    const request: PatchDigitalTwinRequest = {
-      id: {
-        filter: {
-          oneofKind: 'digitalTwin',
-          digitalTwin,
-        },
-      },
-      adminToken: '',
-      forceDelete,
-      operations,
-    };
-
-    return new Promise((resolve, reject) => {
-      this.client.patchDigitalTwin(request, (err, response) => {
-        if (err) reject(err);
-        else if (!response)
-          reject(new SdkError(SdkErrorCode.SDK_CODE_1, 'No patch operation result response'));
-        else {
-          resolve(response.result[0]);
-        }
-      });
-    });
-  }
-
-  /**
-   * @since 0.4.1
-   * @example
-   * // Add a new 'email' property with 'user@example.com' value
-   * async function example(sdk: IdentityClient, accessToken: string) {
-   *   const opResult = await sdk.patchDigitalTwinByToken(accessToken, [
-   *     {
-   *       operation: {
-   *         oneofKind: 'add',
-   *         add: {
-   *           id: '',
-   *           definition: {
-   *             property: 'email',
-   *             context: '',
-   *             type: '',
-   *           },
-   *           value: {
-   *             oneofKind: 'objectValue',
-   *             objectValue: {
-   *               value: { oneofKind: 'stringValue', stringValue: 'user@example.com' },
-   *             },
-   *           },
-   *         },
-   *       },
-   *     },
-   *   ]);
-   *   console.log(JSON.stringify(opResult, null, 2));
-   * }
-   */
-  patchDigitalTwinByToken(
-    token: string,
-    operations: PropertyBatchOperation[],
-    forceDelete = false,
-  ): Promise<BatchOperationResult> {
-    const request: PatchDigitalTwinRequest = {
-      id: {
-        filter: {
-          oneofKind: 'accessToken',
-          accessToken: token,
-        },
-      },
-      adminToken: '',
-      forceDelete,
-      operations,
-    };
-
-    return new Promise((resolve, reject) => {
-      this.client.patchDigitalTwin(request, (err, response) => {
-        if (err) reject(err);
-        else if (!response)
-          reject(new SdkError(SdkErrorCode.SDK_CODE_1, 'No patch operation result response'));
-        else {
-          resolve(response.result[0]);
-        }
-      });
-    });
-  }
-
-  /**
-   * This function deletes the digital twin.
-   * @since 0.4.1
-   * @example
-   * async function example(sdk: IdentityClient) {
-   *   await sdk.deleteDigitalTwin({
-   *     id: 'gid:digitaltwin-id',
-   *     tenantId: 'gid:tenant-id',
-   *     state: DigitalTwinState.ACTIVE,
-   *     kind: DigitalTwinKind.PERSON,
-   *     tags: [],
-   *   });
-   * }
-   */
-  deleteDigitalTwin(digitalTwin: DigitalTwin): Promise<DigitalTwin> {
-    const request: DeleteDigitalTwinRequest = {
-      id: {
-        filter: {
-          oneofKind: 'digitalTwin',
-          digitalTwin,
-        },
-      },
-      adminToken: '',
-    };
-
-    return new Promise<DigitalTwin>((resolve, reject) => {
-      this.client.deleteDigitalTwin(request, (err, response) => {
-        if (err) reject(err);
-        else {
-          if (response && response.digitalTwin) {
-            resolve(response.digitalTwin);
-          } else {
-            reject(new SdkError(SdkErrorCode.SDK_CODE_1, 'Missing delete response'));
-          }
-        }
-      });
-    });
-  }
-
-  /**
-   * This function deletes the digital twin.
-   * @since 0.4.1
-   * @example
-   * async function example(sdk: IdentityClient) {
-   *   await sdk.deleteDigitalTwinByToken('access-token');
-   * }
-   */
-  deleteDigitalTwinByToken(token: string): Promise<DigitalTwin> {
-    const request: DeleteDigitalTwinRequest = {
-      id: {
-        filter: {
-          oneofKind: 'accessToken',
-          accessToken: token,
-        },
-      },
-      adminToken: '',
-    };
-
-    return new Promise<DigitalTwin>((resolve, reject) => {
-      this.client.deleteDigitalTwin(request, (err, response) => {
-        if (err) reject(err);
-        else {
-          if (response && response.digitalTwin) {
-            resolve(response.digitalTwin);
-          } else {
-            reject(new SdkError(SdkErrorCode.SDK_CODE_1, 'Missing delete response'));
-          }
         }
       });
     });
@@ -803,163 +391,6 @@ export class IdentityClient {
   /**
    * @since 0.4.1
    * @example
-   * import * as bcrypt from 'bcrypt';
-   *
-   * async function example(sdk: IdentityClient) {
-   *  const saltRounds = 10;
-   *  const myPlaintextPassword = 'password-of-user';
-   *
-   *  const [salt, hash]: string[] = await new Promise((resolve, reject) => {
-   *    bcrypt.genSalt(saltRounds, function (err, salt) {
-   *      if (err) {
-   *        reject(err);
-   *        return;
-   *      }
-   *
-   *      bcrypt.hash(myPlaintextPassword, salt, function (err, hash) {
-   *        if (err) {
-   *          reject(err);
-   *          return;
-   *        }
-   *        resolve([salt, hash]);
-   *      });
-   *    });
-   *  });
-   *
-   *  const password: PasswordCredential = {
-   *    password: {
-   *      hash: {
-   *        passwordHash: Buffer.from(hash),
-   *        salt: Buffer.from(salt),
-   *      },
-   *      oneofKind: 'hash',
-   *    },
-   *    uid: {
-   *      oneofKind: 'email',
-   *      email: {
-   *        email: 'user@example.com',
-   *        verified: true,
-   *      },
-   *    },
-   *  };
-   *
-   *  const result = await sdk.importDigitalTwins(
-   *    [
-   *      {
-   *        id: '',
-   *        tenantId: 'gid:AAAAAxe5-tWaWUufnFqaMnFwsRk',
-   *        kind: DigitalTwinKind.PERSON,
-   *        state: DigitalTwinState.ACTIVE,
-   *        providerUserInfo: [],
-   *        tags: [],
-   *        password,
-   *      },
-   *    ],
-   *    {
-   *      oneofKind: 'bcrypt',
-   *      bcrypt: {},
-   *    },
-   *  );
-   *  console.log(JSON.stringify(result, null, 2));
-   *}
-   */
-  importDigitalTwins(
-    digitalTwins: ImportDigitalTwin[] | Readable,
-    hashAlgorithm: ImportDigitalTwinsRequest['hashAlgorithm'],
-  ): Promise<ImportDigitalTwinsResponse> {
-    if (!Array.isArray(digitalTwins)) {
-      return this.importStreamedDigitalTwins(digitalTwins, hashAlgorithm);
-    }
-
-    return this.importStreamedDigitalTwins(Readable.from(digitalTwins), hashAlgorithm);
-  }
-
-  private importStreamedDigitalTwins(
-    digitalTwinStream: Readable,
-    hashAlgorithm: ImportDigitalTwinsRequest['hashAlgorithm'],
-  ): Promise<ImportDigitalTwinsResponse> {
-    return new Promise((resolve, reject) => {
-      const results: ImportDigitalTwinsResponse['results'] = [];
-
-      const output = streamSplitter(
-        digitalTwinStream,
-        this.client.importDigitalTwins.bind(this.client),
-        1000,
-        (chunks: unknown[]) => {
-          return ImportDigitalTwinsRequest.create({
-            hashAlgorithm,
-            entities: chunks as ImportDigitalTwin[],
-          });
-        },
-        [new IndexFixer('results.index')],
-      );
-
-      output.on('data', (res: ImportDigitalTwinsResponse) => {
-        results.push(...res.results);
-      });
-
-      output.on('end', () => {
-        resolve({ results });
-      });
-
-      output.on('error', (err) => {
-        reject(err);
-      });
-    });
-  }
-
-  /**
-   * RegisterDigitalTwinWithoutCredential creates a DigitalTwin without credentials, but with properties
-   * This is a protected operation and it can be accessed only with valid agent credentials!
-   * @since 0.4.1
-   * @example
-   * async function example(sdk: IdentityClient) {
-   *   const result = sdk.registerDigitalTwinWithoutCredential({
-   *     tenantId: 'gid:tenant-id',
-   *     digitalTwinKind: grpcIdentityModel.DigitalTwinKind.PERSON,
-   *     digitalTwinTags: ['Employer'],
-   *     properties: [
-   *       {
-   *         definition: {
-   *           property: 'email',
-   *           context: '',
-   *           type: '',
-   *         },
-   *         id: 'property-id',
-   *         value: {
-   *           oneofKind: 'objectValue',
-   *           objectValue: {
-   *             value: {
-   *               oneofKind: 'stringValue',
-   *               stringValue: 'employer@example.com',
-   *             },
-   *           },
-   *         },
-   *       },
-   *     ],
-   *     bookmarks: [],
-   *   });
-   *   console.log(JSON.stringify(result, null, 2));
-   * }
-   */
-  registerDigitalTwinWithoutCredential(
-    request: RegisterDigitalTwinWithoutCredentialRequest,
-  ): Promise<RegisterDigitalTwinWithoutCredentialResponse> {
-    return new Promise((resolve, reject) => {
-      this.client.registerDigitalTwinWithoutCredential(request, (err, response) => {
-        if (err) reject(err);
-        else if (!response) {
-          reject(new SdkError(SdkErrorCode.SDK_CODE_1, 'Missing register digital twin response'));
-        } else {
-          resolve(response);
-        }
-      });
-    });
-  }
-
-  /**
-   * @since 0.4.1
-   * @example
    * async function example(sdk: IdentityClient) {
    *   const result = await sdk.createConsent({
    *     piiPrincipalId: 'gid:principal-id',
@@ -1054,27 +485,6 @@ export class IdentityClient {
         .on('error', (err) => {
           reject(err);
         });
-    });
-  }
-
-  private async processGetDigitalTwinRequest(
-    dtId: DigitalTwinIdentifier,
-    properties?: PropertyMask[],
-  ): Promise<GetDigitalTwinResponse> {
-    const request: GetDigitalTwinRequest = {
-      id: dtId,
-      properties: properties ?? [],
-    };
-
-    return new Promise((resolve, reject) => {
-      this.client.getDigitalTwin(request, (err, response) => {
-        if (err) reject(err);
-        else if (!response)
-          reject(new SdkError(SdkErrorCode.SDK_CODE_1, 'No Digital Twin response'));
-        else {
-          resolve(response);
-        }
-      });
     });
   }
 }
