+++ indykite_sdk_node/src/sdk/model/index.ts
@@ -3,7 +3,6 @@ export * from './consent';
 export * from './consent_challenge';
 export * from './digitaltwin';
 export * from './hash_algorithm';
-export * from './import_digitaltwin';
 export * from './invitation';
 export * from './property';
 export * from './token';
